Check that AWS_SESSION_TOKEN is set in our scripts
Add a check to make sure that `eval
"$("$SCRIPT_PATH/dalmatian-assume-infrastructure-role" -i
"$INFRASTRUCTURE_NAME" -p "$AWS_PROFILE")"` sets AWS_SESSION_TOKEN

`eval` does not error if it is passed nothing to evaluate which can happen if the
the script being called in `$()` fails. `eval` will only pass up an error code if
if it used like `eval /path/to/script`. This form of calling however doesn't
then run export to set the AWS environment variables that the rest of the
script needs to be set.

The `dalmatian-assume-infrastructure-role` script can fail if dalmatian-mfa
hasn't been run recently or the clone of the dalmatian-config repo fails.
We add checking those as suggestions for the user.