Apply linear back-off to AWS API in certificate create
Previously, when we create an SSL cert, we waited a fixed
number of seconds (2) for the AWS API to become consistent.
In practice, we sometimes need to wait longer.

This commit replaces the fixed sleep with a linear back-off.
We assume that the API will become consistent at some point,
and so we don't attempt to break the loop if it does not.
However, If the user gives up and kills the script, we still
attempt to print some useful output.