Add a utility to generate passwords for basic auth
For example:

    ❯ dalmatian util generate-four-words
    Please note that the phrases generated here should not be used as login
    passwords or to hide secrets. Please use 1Password in those cicumstances.
    If you have any questions, please ask the Technical Operations team for advice.

    loves-dara-almost-foreign

Or for use in a script:

    ❯ dalmatian util generate-four-words -q
    protein-oven-placed-mermaid

The corpus of English words is adapted from a shortened version of the 50k
list of common English words used open subtitles, published under CC BY-SA.
The original data and license details can be found here:

    https://github.com/hermitdave/FrequencyWords/

Only words that are between 4 and 7 characters long have been used. The
corpus has been cleaned for profanity, punctuation, numerals, etc.