Add list-services-by-buildspec command
Add a new command to list services in the infrastructure by
buildspec. For example, to list all WordPress services:

    dalmatian config list-services-by-buildspec -b dalmatian_core_buildspec_saluki

Optionally, services can be listed for a single infrastructure:

    dalmatian config list-services-by-buildspec -b dalmatian_core_buildspec_saluki -i dxw-govpress

Note that this command cannot be used to list services which do
not have a buildspec.

Output appears in the following format:

    infra service\n...

e.g

    bbb bbb-fh-next
    dfe-fh web
    dfe-skills web
    dxw-govpress advisories
    dxw-govpress af-covenant
    ...