Add a script to delete unused SSL certificates
'dalmatian certificate delete' will either delete a certificate
with a given ARN ('-c') or all un-issued certificates for
a given domain ('-D').

This is mainly intended to be used in situations where certs
have been created, but could not be vefified, for example due
to missing DNS records.

The '-d' switch can be used to perform a dry-run, if the user
is unsure about what operations the script will perform.

Usage:

    ❯ dalmatian certificate delete -h
    Delete an unused CloudFront certifcate
    e.g.    dalmatian certificate delete -i infra -D example.com
            dalmatian certificate delete -i infra -c arn:aws:acm:region:account:certificate/12345678-1234-1234-1234-123456789012
    Usage: delete [OPTIONS]
      -h                         - help
      -i <infrastructure>        - infrastructure name
      -c <certificate arn>       - remove a single certificate with a given ARN
      -D <domain>                - remove all certificates for domain which are not ISSUED
      -d                         - perform a dry run, do not delete any certificates