List available ec2 instance ids
This script is missing the option to list available ec2 instance IDs which a
user would need if they want to connect manually to a specific instance. For
example if the initial instance connection fails then the user can try
connecting to another one by using an ec2 instance ID from the list.

https://trello.com/c/q6Lomnln/350-dalmatian-command-to-list-ec2-instances