Add dalmatian setup command
This is the beggining of the Dalmatian refactor, which will:
* Use AWS SSO for authentication
* Allow account bootstrapping (using the https://github.com/dxw/terraform-dxw-dalmatian-account-bootstrap
  module)
* Allow deploying to dalmatian rather than through the dalmatian repo
  (https://github.com/dxw/dalmatian)

The setup command configures the required AWS configuration, and stores
it within '~/.config/dalmatian'.

It also sets up the main Dalmatian account and bootstraps it.

The main dalmatian account is the account that holds the S3 bucket for
all the TF Var files.

The resulting '~/.config/dalmatian/setup.json' file can be shared with
other users of the same Dalmatian Project, which can then be used as
defaults by running `dalmatian setup -f setup.json`. It can also be
hosted somewhere, and used by running `dalmatian setup -u
http://example.com/dalmatian-setup.json`