Add commands to trigger SQL backups of an RDS instance/cluster to s3
If we want to trigger a full backup of an RDS instance/cluster to s3 we can do
so with these new commands. This is useful for testing the backup process, or
if you are about to do a risky operation and want to make sure you have a
recent backup first (e.g. a schema change) but don't want to have a local copy
of the database using export-dump instead.

This will be used to trigger a backup of the RDS used by mind sbs when the
migration happens.